节点的指标事件返回准确的Transport的请求数(正在发送+待应答），而不是Channel的待应答请求数

+++ joyrpc-core/src/main/java/io/joyrpc/transport/DecoratorClient.java
@@ -242,7 +242,7 @@ public class DecoratorClient<T extends ClientTransport> implements Client {
     }
 
     @Override
-    public long getRequests() {
+    public int getRequests() {
         return transport.getRequests();
     }
 }
